Add Java native to new smoke_test layout
