--- .github/workflows/test-pull-request-java-native-image.yml
@@ -1,61 +1,0 @@
-name: Test Pull Request (Java Native Image)
-
-on:
-  merge_group:
-  pull_request:
-    branches:
-    - main
-    paths:
-    - 'java/native-image/**'
-
-jobs:
-  smoke:
-    name: Smoke Tests
-    runs-on: ubuntu-latest
-    steps:
-    - name: Setup Go
-      uses: actions/setup-go@v4
-      with:
-        go-version: 1.18.x
-
-    - name: Checkout
-      uses: actions/checkout@v4
-
-    - name: Cache local Gradle repository
-      uses: actions/cache@v3
-      with:
-        path: |
-          ~/.gradle/caches
-          ~/.gradle/wrapper
-        key: ${{ runner.os }}-gradle-${{ hashFiles('java/**/*.gradle*', 'java/**/gradle-wrapper.properties') }}
-        restore-keys: |
-          ${{ runner.os }}-gradle-
-
-    - name: Cache local Maven repository
-      uses: actions/cache@v3
-      with:
-        path: |
-          ~/.m2/repository
-          ~/.m2/wrapper
-        key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml', 'java/**/maven-wrapper.properties') }}
-        restore-keys: |
-          ${{ runner.os }}-maven-
-
-    - name: Setup Directories
-      run: |
-        #!/usr/bin/env bash
-        set -euo pipefail
-
-        mkdir -p ~/.m2/repository
-        mkdir -p ~/.m2/repository/wrapper/dists
-        mkdir -p ~/.gradle/caches
-        mkdir -p ~/.gradle/wrapper/dists
-        chmod -R 775 ~/.m2
-        chmod -R 775 ~/.gradle
-
-    - name: Test Java Native Image Samples
-      run: |
-        ./scripts/smoke.sh --suite java/native-image \
-                           --builder paketobuildpacks/builder-jammy-full:latest \
-                           --builder paketobuildpacks/builder-jammy-base:latest \
-                           --builder paketobuildpacks/builder-jammy-tiny:latest
