+++ java/native-image/public-static-main-native-image-maven/.gitignore
