+++ java/native-image/public-static-main-native-image-maven/pom.xml
