+++ java/native-image/public-static-main-native-image-maven/src/main/java/io/paketo/App.java
