+++ java/native-image/quarkus-native-image-maven/pom.xml
