+++ java/native-image/quarkus-native-image-maven/src/main/java/org/acme/getting/started/GreetingService.java
