+++ java/native-image/quarkus-native-image-maven/src/main/resources/META-INF/resources/index.html
