+++ java/native-image/quarkus-native-image-maven/src/main/resources/application.properties
