+++ java/native-image/quarkus-native-image-maven/src/test/java/org/acme/getting/started/NativeGreetingResourceIT.java
