+++ java/native-image/spring-boot-native-image-gradle/settings.gradle
