+++ java/native-image/spring-boot-native-image-gradle/src/main/java/io/paketo/demo/DemoApplication.java
