+++ java/native-image/spring-boot-native-image-gradle/src/main/resources/application.properties
