+++ java/native-image/spring-boot-native-image-gradle/src/test/java/io/paketo/demo/DemoApplicationTests.java
