+++ java/native-image/spring-boot-native-image-maven/pom.xml
