+++ java/native-image/spring-boot-native-image-maven/src/main/java/io/paketo/demo/DemoApplication.java
