+++ java/native-image/spring-boot-native-image-maven/src/main/resources/application.properties
