+++ java/native-image/spring-boot-native-image-maven/src/test/java/io/paketo/demo/DemoApplicationTests.java
